more *syntax/loc
I was using match-let and got a syntax error that pointed to this file. After changing the match-let definition to use syntax/loc the error pointed to the exact spot causing the problem. Yay!

I changed quite a few syntax-quotes to the */loc form... perhaps some do not need to? I'm not sure.

I also haven't run any tests on this yet.